Restructuring the top-level directory for clarity.
