+++ lsystem/pointer_class.py
